[nick/#81] mission/프로젝트 구조 변경
