+++ Nick/Nike_compose/app/src/main/java/com/clone/nike_compose/presentation/home/HomeScreen.kt
@@ -1,4 +1,4 @@
-package com.clone.nike_compose.ui.home
+package com.clone.nike_compose.presentation.ui.home
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Column
@@ -6,17 +6,14 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.clone.nike_compose.R
-import com.clone.nike_compose.ui.theme.Nike_composeTheme
 
 @Preview(showBackground = true)
 @Composable
