+++ Nick/Nike_compose/app/src/main/java/com/clone/nike_compose/presentation/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.clone.nike_compose.ui.theme
+package com.clone.nike_compose.presentation.ui.theme
 
 import androidx.compose.ui.graphics.Color
